sheet: rework rams so damage dice scaling occurs when added to the sheet
Previously we registered a property that could be accessed from a roll
formula. This worked for basic rolling but failed with some popular roll
enhancement modules on critical hits due to the timing in which this
property was parsed and converted to a value.

The sheet now parses the damage formula and scales the number of dice
(based on ship size) when the collide/ram item is added to the sheet.
This works in all roll enhancement cases.